pyproject.toml
[project]
name = "telibot"
version = "0.1.0"
description = "Telibot, a discord bot!"
authors = ["TeliDev <hiagobvt@live.com>"]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py[voice]>=2.5.2",
    "pydantic-settings>=2.8.1",
    "setuptools>=77.0.3",
]

[tool.pylint]
disable = [
"line-too-long"
]